Add named colours
