+++ package.json
@@ -196,6 +196,7 @@
     "hoist-non-react-statics": "^3.3.0",
     "hsluv": "^0.0.3",
     "lodash": "^4.17.4",
+    "tinycolor2": "^1.4.1",
     "tslib": "^1.9.3"
   }
 }
